Добавлена функция расчёта wastes

+++ blokhina_va/task1/alice/alice_monthly_payment.c
@@ -1,4 +1,4 @@
-extern int credit;
+extern float credit;
 extern float credit_rate;
 extern int months;
